Updated file structure, inheritance, and comments
