--- data/Dialogue/inputdata.h
@@ -1,17 +1,0 @@
-#ifndef INPUTDATA_H
-#define INPUTDATA_H
-
-#include <utility>
-#include <string>
-#include <list>
-#include "fielddata.h"
-
-using namespace std;
-
-class InputData : protected FieldData
-{
-public:
-    InputData();
-};
-
-#endif // INPUTDATA_H
