--- data/Dialogue/mainfielddata.cpp
@@ -1,6 +1,0 @@
-#include "mainfielddata.h"
-
-MainFieldData::MainFieldData()
-{
-
-}
