--- data/Dialogue/subfielddata.h
@@ -1,12 +1,0 @@
-#ifndef SUBFIELDDATA_H
-#define SUBFIELDDATA_H
-
-#include "fielddata.h"
-
-class SubFieldData : protected FieldData
-{
-public:
-    SubFieldData();
-};
-
-#endif // SUBFIELDDATA_H
