--- data/Dialogue/textdata.h
@@ -1,16 +1,0 @@
-#ifndef TEXTDATA_H
-#define TEXTDATA_H
-
-#include "mainfielddata.h"
-
-class TextData : protected MainFieldData
-{
-
-public:
-    // constructors
-    TextData();
-    // deconstructor
-    ~TextData();
-
-};
-#endif // TEXTDATA_H
