+++ data/Fields/MainFields/mainfielddata.cpp
@@ -1,0 +1,16 @@
+#include "mainfielddata.h"
+
+MainFieldData::MainFieldData()
+{
+
+}
+
+bool MainFieldData::getIsAuto()
+{
+    return this->isAuto;
+}
+
+void MainFieldData::setIsAuto(bool value)
+{
+    this->isAuto = value;
+}
