+++ data/Fields/SubFields/subfielddata.h
@@ -1,0 +1,17 @@
+#ifndef SUBFIELDDATA_H
+#define SUBFIELDDATA_H
+
+#include "data/Fields/fielddata.h"
+
+/// <summary>
+/// This class extends FieldData, by...
+/// </summary>
+
+// This class is currently a stub, not sure what else it needs over the base class.
+class SubFieldData : protected FieldData
+{
+public:
+    SubFieldData();
+};
+
+#endif // SUBFIELDDATA_H
